Switch to `kakoune-unwrapped` instead of `kakoune`
The wrapped version of kakoune seems override the PATH variable
without adding the user's PATH back causing kakoune to fail to find
kak-lsp, formatters, and other binaries that are inside my kakrc but
aren't necessary for kakoune to run.

Issue was introduced in this commit:
https://github.com/NixOS/nixpkgs/commit/e4da18a6923e3a332448d96fc03a78feff3b4492

Since i'm not using any features of the wrapped version of kakoune
(plugins), just switch to the unwrapped version.